+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Edited Gitleaks to use our own repo, which removed the deprecated `set-output` command.
+- Enable Renovate to access the repo by default as part of `devctl repo setup`
 
 ## [6.13.0] - 2023-10-05